Initial commit for xp branch: Fixed compile errors with help from Neil Mayhew Fixed a few run-time errors, including an inverted return value on CDialogBar::HandleInitDialog(). Also, the FormantTracker::BuildTrack() had an index out of range problem (not sure why this wasn't a problem before?!)

+++ Src/SaExe/Sa_g_spg.cpp
@@ -524,7 +524,7 @@ BOOL CPlotSpectrogram::OnDrawSpectrogram(CDC *pDC, CRect rWnd, CRect rClip, CSaV
     dlog = new double[nSpectrumIndex + 1];
     for (; nSpectrumIndex>0; nSpectrumIndex--)
     {
-      dlog[nSpectrumIndex] = log(nSpectrumIndex);
+      dlog[nSpectrumIndex] = log(double(nSpectrumIndex));
     }
     dlog[0] = dlog[1];
   }
